Proposal to move K'Rakoth ruins from oceans to arctic planets
I've been thinking about how the player is exposed to the K'Rakoths within the mod and had the thought that maybe it would be more interesting for the K'Rakothan ruins to no longer show up on regular ocean planets, but on Arctic ones instead. This would make the K'Rakoths revealed in a more gradual way, with tier 5 being the first full exposure to them rather than tier 3, which I think would help add a more lovecraftian vibe to them. Plus it'd make Arctic planets more worthwhile to explore with your mod alone.

Obviously this may require adjusting the descriptions of the Arctic planet pin, as well as some adjustments to the PR patch, but I've made sure to ensure compatibility with the Orcana mod, and any others that may add dungeons to Arctic planets.